fix: 스레드 작성자가 본인일 경우에만 수정, 삭제 버튼 노출 및 기능이 동작하도록  구현

+++ backend/books/views.py
@@ -60,9 +60,13 @@ def thread_detail(request, thread_id):
         serializer = ThreadDetailSerializer(thread)
         return Response(serializer.data)
     elif request.method == "DELETE":
+        if thread.user != request.user:
+            return Response({"detail": "권한이 없습니다."}, status=status.HTTP_403_FORBIDDEN)
         thread.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
     elif request.method == "PUT":
+        if thread.user != request.user:
+            return Response({"detail": "권한이 없습니다."}, status=status.HTTP_403_FORBIDDEN)
         serializer = ThreadSerializer(thread, data=request.data, partial=True)
         if serializer.is_valid(raise_exception=True):
             serializer.save()
